# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * The Social Media Room API - private dual-user digital sanctuary
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  /** System login credential — never shown in chat UI */
  email: string;
  /** Preferred display name shown everywhere in the UI */
  displayName: string;
  /** False until the user completes the onboarding form */
  isProfileComplete: boolean;
  /** @nullable */
  avatarUrl?: string | null;
  isAdmin: boolean;
}

export interface ProfileRecord {
  userId: string;
  displayName: string;
  isProfileComplete: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProfileUpdateInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  displayName: string;
}

export type TagType = typeof TagType[keyof typeof TagType];


export const TagType = {
  postbox: 'postbox',
  chat: 'chat',
} as const;

export interface Tag {
  id: number;
  spaceId: number;
  name: string;
  slug: string;
  type: TagType;
  /** @nullable */
  icon?: string | null;
  sortOrder: number;
  unreadCount: number;
}

export interface Space {
  id: number;
  name: string;
  sortOrder: number;
  tags: Tag[];
}

export interface SpaceInput {
  /** @minLength 1 */
  name: string;
  sortOrder?: number;
}

export interface SpaceUpdateInput {
  /** @minLength 1 */
  name?: string;
  sortOrder?: number;
}

export type TagInputType = typeof TagInputType[keyof typeof TagInputType];


export const TagInputType = {
  postbox: 'postbox',
  chat: 'chat',
} as const;

export interface TagInput {
  spaceId: number;
  /** @minLength 1 */
  name: string;
  type: TagInputType;
  icon?: string;
  sortOrder?: number;
}

export interface TagUpdateInput {
  /** @minLength 1 */
  name?: string;
  icon?: string;
  sortOrder?: number;
  spaceId?: number;
}

export type UnreadCountsCounts = {[key: string]: number};

export interface UnreadCounts {
  counts: UnreadCountsCounts;
}

export interface LetterSummary {
  id: number;
  title: string;
  tagIds: number[];
  authorId: string;
  authorName: string;
  createdAt: string;
  isRead: boolean;
  hasAttachments: boolean;
  /** @nullable */
  excerpt?: string | null;
}

/**
 * Editor.js-style block content JSON
 */
export type LetterContent = { [key: string]: unknown };

export type LetterReactions = {[key: string]: string[]};

export interface Attachment {
  id: number;
  url: string;
  filename: string;
  mimeType: string;
  size: number;
}

export interface Letter {
  id: number;
  title: string;
  /** Editor.js-style block content JSON */
  content: LetterContent;
  tagIds: number[];
  authorId: string;
  authorName: string;
  createdAt: string;
  isRead: boolean;
  reactions: LetterReactions;
  attachments: Attachment[];
}

export type LetterInputContent = { [key: string]: unknown };

export interface AttachmentInput {
  url: string;
  filename: string;
  mimeType: string;
  size: number;
}

export interface LetterInput {
  /** @minLength 1 */
  title: string;
  content: LetterInputContent;
  /** @minItems 1 */
  tagIds: number[];
  attachments?: AttachmentInput[];
}

export interface ReactionInput {
  emoji: string;
}

/**
 * Map of emoji to list of user IDs who reacted
 */
export type MessageReactions = {[key: string]: string[]};

export interface Message {
  id: number;
  tagId: number;
  /** Unique Clerk user ID of the sender */
  senderId: string;
  /** Preferred display name shown in the chat UI for both parties */
  senderDisplayName: string;
  /**
     * System login email stored for auditing — never rendered in the UI
     * @nullable
     */
  senderLoginName?: string | null;
  content: string;
  /** Server-authoritative UTC timestamp */
  createdAt: string;
  /** True once the other user has loaded messages up to and including this one. Once true, the sender can no longer delete it. */
  seenByPartner: boolean;
  /** Map of emoji to list of user IDs who reacted */
  reactions: MessageReactions;
}

export interface MessageInput {
  tagId: number;
  /** @minLength 1 */
  content: string;
}

export interface MessageUpdateInput {
  /** @minLength 1 */
  content: string;
}

export type ReactionToggleResultReactions = {[key: string]: string[]};

export interface ReactionToggleResult {
  reactions: ReactionToggleResultReactions;
}

export interface LetterComment {
  id: number;
  letterId: number;
  userId: string;
  authorName: string;
  content: string;
  createdAt: string;
}

export interface LetterCommentInput {
  /** @minLength 1 */
  content: string;
}

export interface NotificationSummary {
  /** Count of new emoji reactions on your messages since last seen */
  newMessageReactions: number;
  /** Count of new comments on your letters since last seen */
  newLetterComments: number;
  /** Total emoji reactions on your letters from the partner (all time) */
  totalLetterReactions: number;
  /** The timestamp used as the "since" cursor */
  since: string;
}

export interface UploadResult {
  url: string;
  filename: string;
  mimeType: string;
  size: number;
}

export type GetLettersParams = {
tagId: number;
};

export type GetMessagesParams = {
tagId: number;
/**
 * Cursor - get messages before this message ID
 */
before?: number;
limit?: number;
};