# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Social Media Room API - private dual-user digital sanctuary
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /letters:
    get:
      operationId: getLetters
      tags: [letters]
      summary: List letters for a given tag
      parameters:
        - name: tagId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of letters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LetterSummary"
    post:
      operationId: createLetter
      tags: [letters]
      summary: Create a new letter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LetterInput"
      responses:
        "201":
          description: Created letter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Letter"
  /letters/{letterId}:
    get:
      operationId: getLetter
      tags: [letters]
      summary: Get a specific letter (marks as read)
      parameters:
        - name: letterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Letter with full content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Letter"
    delete:
      operationId: deleteLetter
      tags: [letters]
      summary: Delete a letter
      parameters:
        - name: letterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /letters/{letterId}/react:
    post:
      operationId: reactToLetter
      tags: [letters]
      summary: Add or remove an emoji reaction to a letter
      parameters:
        - name: letterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Updated letter reactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Letter"
...
  /letters/{letterId}/comments:
    get:
      operationId: getLetterComments
      tags: [letters]
      summary: Get all comments for a letter
      parameters:
        - name: letterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LetterComment"
    post:
      operationId: addLetterComment
      tags: [letters]
      summary: Add a comment to a letter
      parameters:
        - name: letterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LetterCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LetterComment"
...
components:
  schemas:
...
    LetterSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        tagIds:
          type: array
          items:
            type: integer
        authorId:
          type: string
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time
        isRead:
          type: boolean
        hasAttachments:
          type: boolean
        excerpt:
          type: ["string", "null"]
      required:
        - id
        - title
        - tagIds
        - authorId
        - authorName
        - createdAt
        - isRead
        - hasAttachments
    Letter:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: object
          description: Editor.js-style block content JSON
          additionalProperties: true
        tagIds:
          type: array
          items:
            type: integer
        authorId:
          type: string
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time
        isRead:
          type: boolean
        reactions:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
      required:
        - id
        - title
        - content
        - tagIds
        - authorId
        - authorName
        - createdAt
        - isRead
        - reactions
        - attachments
    LetterInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: object
          additionalProperties: true
        tagIds:
          type: array
          items:
            type: integer
          minItems: 1
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentInput"
      required:
        - title
        - content
        - tagIds
    Attachment:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        filename:
          type: string
        mimeType:
          type: string
        size:
          type: integer
      required:
        - id
        - url
        - filename
        - mimeType
        - size
    AttachmentInput:
      type: object
      properties:
        url:
          type: string
        filename:
          type: string
        mimeType:
          type: string
        size:
          type: integer
      required:
        - url
        - filename
        - mimeType
        - size
    ReactionInput:
      type: object
      properties:
        emoji:
          type: string
      required:
        - emoji
...
    LetterComment:
      type: object
      properties:
        id:
          type: integer
        letterId:
          type: integer
        userId:
          type: string
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - letterId
        - userId
        - authorName
        - content
        - createdAt
    LetterCommentInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required:
        - content
...