# lib/api-zod/src/generated/types/letterComment.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * The Social Media Room API - private dual-user digital sanctuary
 * OpenAPI spec version: 0.1.0
 */

export interface LetterComment {
  id: number;
  letterId: number;
  userId: string;
  authorName: string;
  content: string;
  createdAt: Date;
}